improve testing

+++ catkit/flow/tests/test_qeio.py
@@ -1,5 +1,6 @@
+from catkit.flow import qeio
+from ase.build import bulk
 import unittest
-from catkit.flow.qeio import cd
 import shutil
 import os
 
@@ -9,12 +10,20 @@ class TestQEIO(unittest.TestCase):
 
     def tearDown(self):
         """Clear the temp file."""
-        shutil.rmtree('temp')
+        if os.path.exists('temp'):
+            shutil.rmtree('temp')
 
     def test_cd(self):
         """Test the cd context manager."""
-        with cd('temp'):
+        with qeio.cd('temp'):
             assert(os.getcwd().split('/')[-1] == 'temp')
+
+    def test_geometry_hash(self):
+        """Test the geometry_hash function."""
+        atoms = bulk('Pd', cubic=True)
+        ghash = qeio.geometry_hash(atoms)
+
+        assert(ghash == '2980adf7416f0035b803eabeef085ecf')
 
 
 if __name__ == '__main__':
